Use nanoTime to compute delay time since first attempt
This was (incorrectly) changed in the re-retrying fork to use
currentTimeMillis() instead of nanoTime() to make the delay computation
"simpler", but it resulted in a potentially problematic calculation
since nanoTime is intended to be used for elapsed time computations in
Java. Using currentTimeMillis can can odd issues (e.g. a call to
currentTimeMillis could return a number less than a preceding call
and thus result in a negative elapsed time)

Fixes #13