Sonar: Make Attempt transient in RetryException
Mark the Attempt field in RetryException as transient to make Sonar
happy.

While technically Sonar rule java:S1948 (Fields in a "Serializable"
class should either be transient or serializable) is true, it seems
highly unlikely that an exception would be serialized to disk during
any processing. The rule gives the example of a JEE application server
flushing objects to disk under heavy load, but again it seems very
unlikely that an exception would ever be serialized, since exceptions
generally need to be handled (or simply propagated) "right now".

If I am wrong about this behavior and a RetryException ever gets
serialized to disk, and then gets de-serialized somewhere else, and then
someone attempts (ha, a pun) to call getNumberOfFailedAttempts(), they
could get a NPE. While not lovely, I'm not going to (at present unless
someone shows me a real-life example of this happening) modify the
getNumberOfFailedAttempts() method to handle this case other than to
check that the lastFailedAttempt is non-null, and throw an
IllegalStateException if it actually is null. So they will receive
an IllegalStateException with an explanation instead of an NPE, which
ultimately I suppose is better.